Added the script tag. Added the click event for the top navigation bar.
Added the JavaScript script tag to the head of the document with the attribute defer.
Added the script for the top navigation bar to toggle visible and hide when the hamburger menu is clicked.
Added overflow-x hidden to the CSS body selector to stop the horizontal overflow.
Added another CSS selector to hide the navigation bar when toggled with JavaScript.