Added a margin to the portfolio website slider container figure element. Added a margin to the left and right arrows. Added a border radius to the image slider. Added a comment containing the media query for the future. Added a temporary background comment
Added a top margin to the portfolio website slider container figure element.
Added a top margin and small left and right margin to the left and right arrows.
Added a border radius to the image slider.
Added a comment containing the media query for the future.
Added a temporary background comment on the image slider.