Noticed a spelling mistake in the landing page title, Added a class to the portfolio website content container. Added an article element to the website content. Changed the form title values. Changed the message form input element to a text area element.
Noticed a spelling mistake in the landing page title and changed it to the correct spelling,
Added a class to the portfolio website content container for the accordion.
Added an article element to the website content and gave it a class for the show more accordion arrow.
Changed the form title values.
Changed the message form input element to a text area element.