Changed the margins on the portfolio website container.
Added a class to the top hr element so it is separate, that way it will not affect the other elements on top or underneath the portfolio websites when I add more in the future.
Changed the margins for the portfolio website slider and title so it is less code and more maintainable in the future.