Changed the path coordinates for Chrome compatibility. Changed the label and input to text instead of phone. Added a label animation to the form. Changed the background colour for the profile container. Added a background colour to the portfolio website container. Changed the width of the portfolio website container. Added border radius to the portfolio website container and contact form.  Changed the styles for the contact form.
Changed the path coordinates for Chrome compatibility as it would not render the paths with the same coordinates.
Changed the label and input to text instead of phone because the label animation would not work properly when hovering over and clicking on the phone input.
Added a label animation to the form. Changed the background colour for the profile container by adding another linear-gradient RGBA colour to make it show the secondary colour variable more.
Added a background colour to the portfolio website container.
Changed the width of the portfolio website container.
Added border radius to the portfolio website container and contact form.
Changed the styles for the contact form such as the caret colour, user-select and label font colour.