Added target blanks and rel noopener to some of the links. Changed the text for the portfolio website titles. Removed all temporary borders. Added a bottom padding to the form container.
Added target blanks and rel noopener to some of the links this allows some of the links to open on a separate page and stop malicious attacks on the links.
Changed the text for the portfolio website titles.
Removed all temporary borders.
Added a bottom padding to the form container.