Changed the SVG laptop image and made the bottom width smaller and emphasized the keys. Added media queries to the layout for smaller screens. Added spacing to the profile section to make the design look clearer to the user. Changed the width of the portfolio section so it fits better on smaller screens.
Changed the SVG laptop image and made the bottom width smaller and emphasized the keys to make them stand out more to the user.
Added media queries to the layout for smaller screens to make the design look clearer for the user.
Added spacing to the profile section to make the design look clearer to the user.
Changed the width of the portfolio section so it fits better on smaller screens instead of it being too small for the user.