Removed an unnecessary class. Added the imaged to the second image slider. Separated the copyright year and sentence in order to make the year update automatically using JavaScript. Added the script to update the copyright year. Changed the size of the icons on the navigation bar and footer. Added a hover affect to the footer icons.
Removed an unnecessary class.
Added the imaged to the second image slider.
Separated the copyright year and sentence in order to make the year update automatically using JavaScript.
Added the script to update the copyright year.
Changed the size of the icons on the navigation bar and footer.
Added a hover affect to the footer icons which changed the colour on hover.