Added new image templates to the image slider. Removed the old temporary images for the image slider. Added a new class to the welcome page title to add a user selection of none. Added highlighted text to the portfolio introduction, show more accordion and contact introduction. Added the icons and text to the second technologies used container. Added cursor properties to the CSS. Added user-select none to some of the elements on the page. Moved some of the CSS to make it cleaner. Added transitions to the show more arrow and links. Added text shadows to the section introduction titles and portfolio website highlighted text. Added an inline style to the portfolio introduction text. Added the script to make the website have smooth scroll compatible for all browsers as well as old versions.
Added new image templates to the image slider.
Removed the old temporary images for the image slider.
Added a new class to the welcome page title to add a user selection of none.
Added highlighted text to the portfolio introduction, show more accordion and contact introduction to make it stand out more.
Added the icons and text to the second technologies used container. Added cursor properties to the CSS for any element that is a link or button for a better user experience.
Added user-select none to some of the elements on the page.
Moved some of the CSS to make it cleaner.
Added transitions to the show more arrow, buttons and links.
Added text shadows to the section introduction titles and portfolio website highlighted text.
Added an inline style to the portfolio introduction text.
Added the script to make the website have smooth scroll compatible for all browsers as well as old versions.