Added form validation and changed validation message.
Added the name attribute to the input element.
Changed the Full Name form validation message.
Added the form validation class selectors in the CSS, I had to make the properties the same as the .form-input class because the style would change when the function was triggered.
Added the form validation script.