Updated the form validation
Set Autocomplete to off for UI purposes.
Added separate classes to the form paragraphs.
Set the form paragraphs to visibility hidden to stop them being visible when the page loads.
Changed the script for the form validation to stop the form paragraphs from being visible, when the validation becomes valid and the text is deleted the form fields.