Updated contact form
Made the submit button functional using FormSubmit in the script, it allows the form to only send when all validation is valid and disappears when the page refreshes or the text in the input field is deleted.
Added a template input element so when the form goes to my email it has a template to make the form more visible to read.
Added a subject input element to add text to the subject of the email.
Added a custom thank you page input element.