Added content to the thank you page
Added the top navigation bar to the HTML, CSS and JavaScript files to make it the same as the portfolio top navigation bar.
Added the links in the HTML to the CSS and JavaScript files.
Added the semantic elements to the HTML for the thank you animation, paragraphs, envelope animation and buttons.
Added @font-face for the fonts from font-awesome to the CSS.
Styled and positioned the thank you animation, paragraphs, envelope animation and buttons.