Updated SVG images and fixed navigation bar.
Fixed the navigation bar disappearing on click at the responsive size of 410px or less, by adding the CSS property of right and a value of 0 to the class accordion navigation bar hide.
Added a gradient to the thank you message underline.
Changed the colours of the envelope SVG image to make the thank you paragraph stand out more.
Changed the width of the button containers and adjusted the media query width.