Added animation and links to the navigation.
Added the animation for the thank you message.
Added an icon to the thank you page.
Added the links to the hypertext ref attributes on the thank you page.
Added the width and height of the thank you message SVG image to the external CSS file.
Added a pointer cursor and filter to the button hover.