Added a dark mode pop up
I gave it a lot of thought and decided to ask the user for permission to save their data to local storage when clicking on the moon icon to allow or decline dark and light mode.
I added a click event to the icon which shows a pop up box when clicked.
I added some of the elements that will change colour when the user allows saved local storage, it is commented out temporarily until I add click events to the allow and decline buttons.