Updated dark mode
Changed the pop up box user interface.
Added the script for dark mode, when the user clicks on the moon icon a pop up box appears. It gives the user a choice to allow or decline their preferences to be saved to local storage. If they decline it the pop up box disappears until they click on the moon icon again which will show the pop up box again. If they allow dark mode the pop up box disappears and is saved in local storage, the dark mode classes are then added to the elements which changes the colours of the elements. Each time the user clicks on the moon icon after that, the dark mode classes are removed and toggled between enabled and disabled.