Updated dark mode
Added a comment in the HTML file to make the show more arrow SVG wider in the future.
Added more variables and class lists to the dark mode function for the elements that change colour in dark mode.
Fixed the dark mode issue in the form, users can now type into the input field and the form input background will not change back to light mode.