Updated dark mode on form
Fixed the form to ensure it only sends when all form validation is hidden. The message input does not have the class form-input-message removed when the message is deleted from the message input.
Added a DOMContentLoaded event listener to the form to ensure the function runs when the document is loaded.
Added JavaScript styles for the placeholder pseudo elements for light and dark mode, unfortunately JavaScript does not allow us to select pseudo elements. This was the only way to change the colours.
Made changes to the form when the website is in dark mode to make sure it works the same as when it is in light mode.