Updated dark mode on the thank you page
Added dark mode to the contact form thank you page. If the user has saved their preference on the portfolio, the thank you page loads their preference.
I then added functions for if the user clicks on the moon icon and is then able to toggle from light to dark mode.
Added the dark mode selectors in the CSS file.
Changes some of the class names in the HTML and CSS files.