Updated image slider and added light mode icon
Removed unnecessary images for the image slider.
Tested the image slider using images which I have added temporarily.
In order for the user interface to look better for the users on mobile devices and not lose image quality, I changed the width of the image slider by using media queries and object fit. I also changed the image slider layout for mobile devices so the user can see the image more as the image was being cut.
Added a sun icon to replace the moon icon when the website changed back to light mode.