Updated form. Updated styles.
Spent a good few days fixing the form. Formsubmit.co was not working properly, it was saying that I needed to use method="POST" which I have. I solved the problem by finding another third party website which was able to send forms to email without the need for backend. I found web3forms.com, this website made it much easier to set up the form and provides a hCaptcha that can go underneath the form, rather than a separate page.
I have put the HTML code that I was using from formsubmit into comments, just in case I need it in the future. Once the website goes live and I test the new forms cookies are configured as they should be, I will remove the comments with the old HTML code.
I added the thank you page to the form so that the user is redirected to it, once the form is submitted.
I have added a comment with the privacy policy paragraph for when I create the privacy policy link, I will then take that comment out to add to the new privacy policy page.
I changed the footer links to credit web3forms.com instead of formsubmit.co.
Changed the font size for the sun icon to be the same size as the moon icon.
Changed the styles for the caret on the form, it is visible when the user typed in the input boxes but 'hidden' when the user focuses on the input boxes or there is an empty input field.
Added visibility hidden to the input field placeholders for when the page is first viewed and dark mode has not been set.
Added positioning to the hCaptcha element.
Added the function to set the cookie when the form is valid, the hCaptcha is complete and the user clicks on the submit button.
Added a script for the hCaptcha to change colours between light and dark mode.
Fixed a styling issue in the script for the form labels.
Added a script for the form button to be visible as it was hidden when the message input field became empty.
Changed the scripts maintainability for the valid form inputs and the messages that appear when not valid.
Changed the script for the form submission.
Added the sun icon to the thank you page.
Changed the position of the sun and mood icon to make sure it was in the same position as the portfolio page.
Changed the function for dark and light mode for the sun and mood icon on the thank you page.