Updated privacy policy page
Removed the comment with the hCaptcha privacy policy page from the portfolio HTML file and added it to the privacy policy page HTML file.
Added the privacy policy page link to the footer of the portfolio HTML file.
Added the basic boilerplate to the privacy policy page HTML file.
Added the basic boilerplate to the privacy policy CSS file.
Added the styles for the privacy policy link in the footer of the portfolio in the portfolio CSS file.
Added the JavaScript file and linked it with the HTML file for the privacy policy page.