Updated privacy policy page
Moved a comment for the CSS Variables in the portfolio CSS file.
Added a comment for the CSS Variables in the thank you page CSS file and privacy policy page CSS file.
Added the colour variables to the privacy policy page CSS file in the root element.
Added the fonts from Google Fonts to the privacy policy page CSS file.
Added the boilerplate for the body element in the privacy policy page CSS file.