Updated privacy policy page
Added the rel="noreferrer" attribute to the index.html and thank-you-page.html files for older browsers.
Added the privacy policy information and links to the privacy-policy-page.html file.
Added the styles and font from Google Fonts to the privacy-policy-page.css file.
Added a .forEach() method to the script to the privacy-policy-page.js file for all titles on the privacy-policy-page.html.