Updated privacy policy page
Added an anchor tag for the back to top footer button, this allows the user to scroll back to the top of the page when they scroll down to the footer.
Add the rel="noopener" and rel="noreferrer" attributes for added security.
Added the styles for the dark mode pop up box.
Fixed the footer by removing the z-index property on the footer-content selector, this allows the user to be able to click on the links.
Styled the anchor tag that allows the user to scroll back to the top of the page.
Added a z-index property for the top navigation bar so that it is visible when the user scrolls to the footer.
Added the script for the smooth scroll.
Changed the script for dark mode, which includes the dark mode pop up box if dark mode is not saved on local storage, when the user clicks on the moon icon.