Updated thank you page
Changed the relative link path for the privacy policy link in the footer on the privacy-policy-page.html.
Added the main element to make the thank-you-page.html file have more semantic meaning.
Added the back to top anchor link on the thank-you-page.html file.
Added the footer from the portfolio.html file to the thank-you-page.html file.
Added the styles for the dark mode pop up box in the thank-you-page.css file.
Added the styles for the footer in the thank-you-page.css file.
Added the styles for the back to top link anchor tag in the thank-you-page.css file.
Added/changed the styles for dark mode in the thank-you-page.css file.
Changed the styles for the message sent paragraph in the thank-you-page.css file.
Changed the height of the buttons on the thank-you-page.css file.
Added the styles for the footer in the thank-you-page.css file.
Added the script for smooth scroll in the thank-you-page.js file.
Changed the script for dark mode in the thank-you-page.js file.
Changed the speed of the thank you message SVG animation in the thank-you-page.js file.