Updated CSS and JS files
Added the "use strict" literal expression to the script.js, thank-you-page.js and privacy-policy-page.js files.
Changed the hover styles for the buttons on the style.css and the thank-you-page.css files.
Changed the link styles for the external links on the privacy-policy-page.css file.
Added a landing page media query to the style.css file and made the title change from column to row when the page gets to a max width of 770px.