Updated primary colour variable, selector class names and portfolio section.
Updated the selector class names in the index.html, style.css and script.js files and made them more meaningful.
Updated the portfolio section in the index.html file, I commented out the portfolio website container ready for it to be added to in the future.
Changed the colour of the primary colour variable.
Changed the div element to a header element for the portfolio container to make it more semantic.