Updated HTML class names CSS and selectors
Changed the HTML class names in the index.html file to make them more readable and meaningful.
Changed the CSS selectors in the style.css file.
Changed the selectors in the script.js file.
Removed any unnecessary code and commented out some of the code to check it in the future to make sure it is not needed before removing them.