Updated style.css and script.js files
Commented out the .portfolio-websites-show-more-accordion-container-button > selector as I don't think it is needed and I was able to add a dark mode classList to the show more arrow on the dark mode function. Still need to do more tests to make sure it does not affect anything else on the portfolio.
Changed the variable declaration names on the script.js file to make them more readable and meaningful.