Updated index.html HTML class names, style.css CSS selectors and order of CSS, script.js variable names
Changed the HTML class names in the index.html file to make the code more readable and meaningful.
Changed the CSS selectors in the style.css file.
Changed the order of the media query declaration blocks in the style.css file, from smallest screen, to biggest screen to make it more readable.
Changed the variable names in the script.js file to make the code more readable and meaningful.
Removed the unnecessary code from the script.js file
Changed the order of the dark mode function in the script.js file to make it more readable.