Updated HTML class name and JavaScript variable, parameter, argument and method names
Changed the HTML class name in the index.html file to fix the image slider containers.
Changed the JavaScript variable names in the script.js file to make it more meaningful and readable.
Changed the parameter and argument names in the image slider function in the script.js file to make it more meaningful and readable.