Updated index.html, script.js, thank-you-page.html and thank-you-page-style.css
Changed the meta tag content in the index.html file for SEO purposes.
Reduced repetitive code by creating variables for the dark mode function.
Changed the meta tag order and content in the thank-you-page.html file.
Made a start on changing the class names in the thank-you-page.html file to make them consistent with the index.html file.
Changed the selector names in the thank-you-page-style.css file.