Updated the index.html, thank-you-page.html, thank-you-page-style.css and thank-you-page-script.js files
Added a rel="noopener noreferrer" attribute to the portfolio view website and view source code buttons on the index.html file.
Changed the class names on the thank-you-page.html file to make it more readable and meaningful.
Changed the selectors on the thank-you-page-style.css file.
Removed any unnecessary classes and selectors from the thank-you-page.html and thank-you-page.css files.
Changed the variable selectors in the thank-you-page-script.js file.