Updated the index.html, style.css, script.js, thank-you-page.html, thank-you-page-style.css, thank-you-page-script.js, privacy-policy-page.html, privacy-policy-page.css and privacy-policy-page.js
Changed some of the grammar in the index.html file.
Cleaned up the index.html code.
Changed the order of the media queries in the style.css file in order for it to work.
Changed the dark mode pop up box function in the script.js file, to only allow users to be able to click on the dark mode icon container once and allow the pop up container to be added only the once.
Removed most of the declaration blocks from the thank-you-page-style.css and the privacy-policy-page.css files and linked them to the style.css page, reducing the size of the thank-you-page-style.css and privacy-policy-page.css files and making the pages more maintainable.
Changed the variable names and selectors in the thank-you-page.html and privacy-policy-page.html, thank-you-page-style.css, privacy-policy.css, thank-you-page-script.js and privacy-policy-page.js files to make them more readable.