Updated the index.html, style.css, script.js, thank-you-page-script.js and privacy-policy-page.js files
Changed the portfolio view website and view source code buttons to an anchor element, to improve accessibility and best practices in the index.html file.
Changed the z-index values to put them in numerical order in the style.css file.
Changed the width of the technology used container divides to ensure the flex wrap works as best as it can and allow responsiveness on smaller screen sizes without it breaking in the style.css file.
Changed the footer bottom privacy policy link to float center to make the UI better in the style.css file.
Updated the class names and variable names for the dark mode modal in the style.css, script.js, thank-you-page-style.js and privacy-policy.js files.