Updated script.js contact form
Fixed the bug on the contact form, when the input field is invalid or valid and the dark/light mode container was clicked, the form input would remain the same color as the dark/light mode and not change. I was able to fix this by adding an if statement to the enableDarkMode and disableDarkMode functions.
Commented out the contactFormValid and contactForm Invalid forEach methods on the enableDarkMode and disableDarkMode functions as I believe they are not needed now I have fixed the bug, still need to do more tests to confirm.