Updated index.html, style.css, script.js, thank-you-page.js and privacy-policy-page-script.js files
Added a container to the contact form submit button so that it can be centered in the index.html file.
Changed the order of the CSS modal overlay declaration block for the dark mode and hCaptcha modals in the style.css file.
Adjusted the order of the border size, style and color values in the style.css file.
Added the user-select property and the value of none to the dark mode and hCaptcha modals in the style.css file.
Fixed the width of the download and submit buttons, I have made them responsive and the same size as each other when the screen size changes.
Removed unnecessary code from the script.js file.
Added the overlay for the dark mode modal, to ensure the user is not able to click on anything on the page after the modal appears on the thank-you-page-script.js and privacy-policy-page.js files.