Updated index.html, style.css and privacy-policy-page.css files.
Added the tabindex attribute to the clickable HTML elements ready for the accessibility functions in the index.html files.
Added media orientation landscape styles for the dark mode and hCaptcha modals, making sure the buttons are accessibly on landscape and changing the color of the fonts in the style.css file.
Removed the media query, padding left property for the section title in the privacy-policy-page.css file.