Updated accessibility in the index.html, style.css and script.js files
Changed the dark mode icon container from a nav element to a button element for keyboard accessibility.
Gave the dark mode button an aria-label for accessibility.
Changed the alt description for the profile headshot to something more simple and specific.
Commented out the HTML for the old dark mode container until the end when I test it fully, in case it is needed in the future for whatever reason.
Added the styles for the dark mode button, I added a focus-visible pseudo class to enable WCAG standards and in order to stop the focus border being visible, when the user clicks on the dark mode button using a mouse.
Commented out the CSS for the old dark mode container until the end when I test it fully, in case it is needed in the future for whatever reason.
Changed the dark mode script to have a keydown event so that the user can not only click on the dark mode button, but also be able to use tab to select and toggle the dark mode.
Cleaned up the dark mode script and gave it better indentation.
Commented out the JavaScript for the old dark mode container until the end when I test it fully, in case it is needed in the future for whatever reason.