Updated accessibility in the index.html and style.css files.
Changed the navigation accordion links by adding separate list elements to improve with accessibility in the index.html file.
Added a role attribute the the navigation for added improvement to accessibility in the index.html file.
Added an aria-label for screen readers in the index.html file.
Added an aria-label and a hypertext reference attribute on the anchor tag to improve accessibility in the index.html file.
Styled the navigation accordion so it did not change the original styles in the style.css file.
Changed the selector names in the style.css file.