Updated accessibility in the index.html, style.css and script.js files
Changed aria-label, capitalization in the index.html file.
Changed the aria-describedby value to the dark and light mode icons and added a span with a screen reader only attribute so that screen readers are able to identify what the icons are named to the index.html file.
Changed the aria-label values on the anchor tags which are used to scroll the user to the section when clicked, I made them more descriptive to screen readers, I also added a tabindex of -1 so the elements are not able to use tab on the elements in the index.html file.
Added aria-hidden true to all elements that are used for decorative purposes only in the index.html file.
Changed the width of the landing page welcome backgrounds to make them more aligned with the title content in the style.css file.
Added the selectors for the landing page welcome background dark mode in the style.css file.
Added an event listener to the top of the script to check which elements are being focused when tabbing through the page to check for accessibility issues in the script.js file.
Added to the dark mode function, the landing page title background, color change to make the contrast better for accessibility in the script.js file.