Updated accessibility in the index.html, style.css and script.js files.
Added aria hidden true to the logo and hamburger menu SVG path and circle elements in the index.html file.
Added labels to the anchor elements to improve accessibility in the index.html file.
Added semantic elements for accessibility improvement in the index.html file.
Added screen reader only attributes to span elements for better accessibility in the index.html file.
Changed the image slider by adding a semantic interactive element for accessible users, aria labels and roles to improve accessibility in the index.html file.
Added aria labels, aria live, role and screen readers only to the view website and view source code buttons to improve accessibility in the index.html file.
Added the class selector for screen readers only to the top of the CSS file, this is to ensure that the style is rendered as quick as possible to stop it from showing when the page loads or refreshes in the style.css file.
Changed the image slider styles in the style.css file.
Improved indentation in the script.js file.
Changed the functionality of the image slider when the screen size is less than 600px wide in the script.js file.
Added an announcement for screen readers when the page is opened in a new page for better accessibility in the script.js file.