Updated accessibility in the index.html, style.css and script.js files
Added an aria label to the figure element to improve accessibility.
Removed the tab index attributes from the contact form submit button and it's container as it is redundant.
Changed the footer navigation elements by adding separate list elements to improve accessibility.
Added a list style type property to the footer navigation container and footer bottom left icons.
Changed the function for the contact form to help with accessibility, the submit button is now focusable, however it is not able to be submitted when the user clicks on it or used tab and enter or space, if the contact form is invalid.