Updated accessibility in the index.html, style.css and script.js files
Added aria hidden to the technologies used icons as they are not needed to have a label for screen readers, the figcaption element is sufficient enough for accessibility in the index.html file.
Changed the placeholder values to make them better for accessibility in the index.html file.
Removed the aria describedby attributes and added titles and aria labels directly to the a tag for the LinkedIn and GitHub icons to improve accessibility in the index.html file.
Changed the CSS variable for the secondary colour (I kept the original just in case it is needed in the future) in the style.css file.