Updated accessibility in the index.html, style.css and script.js files
Added a contrast button for the future in the index.html file.
Added separate classes to the profile introduction container paragraphs, so that I could style the profile introduction container greeting separately allowing it to have a background around the text, rather than the whole width of the container in the index.html file.
Styled the contrast button in the style.css file.
Changed the positioning of the dark mode and light mode icons to make them the same size as the contrast button in the style.css file.
Added a darker background to the primary coloured text with a white background to improve accessibility contrasts in the style.css file.
Removed any white space to reduce the file size and readability in the style.css file.