Updated accessibility in the index.html, style.css and script.js files
Added a different class for the (opens in a new tab) text for screen readers and set the same properties as the .sr-only class in the index.html and style.css files.
Removed unnecessary code in the style.css file.
Changed the padding, line height and position of the profile content, quotation container and quotation marks in the style.css file.
Changed the profile content and profile show more accordion contents margin for smaller screens in the style.css file.