Updated accessibility in the index.html and style.css files.
Changed the profile quotation content and added a cite element for more semantic meaning.
Added an aria label attribute to the download CV button to help the user know what the button does and be more specific with the meaning.
Added icons to the buttons to help the user get a visual interpretation of what the buttons do and improve the design of the website.
Added aria hidden true to the ::before pseudo-elements in the portfolio and contact section, this allows visual users to be able to see it whilst not being read out twice on screen readers.
Added a div to each slider image and gave them an attribute of aria live polite and aria-atomic true so when the user clicks on the left and right arrows, the alt attribute is read out to the user on the screen reader.