Updated accessibility in the index.html and script.js files
Commented any unnecessary code.
Changed the name of the aria label for the portfolio website buttons and portfolio websites show more accordion container.
Added an aria live and aria atomic attribute to the Show More text on the portfolio show more accordion container, this is to ensure that the screen reads out live changes to the text.
Added aria hidden to the show more arrow as it is only used for decorative purposes.
Added an aria-expanded to the portfolio websites show more accordion container the screen reader, reads out that it is expanded or collapsed.
Added an aria-label to the form element to make it more clear for users.
Added spaces to the form for readability and the future accessibility features that I will be adding in the future.
Added the dynamic change to the aria-expended attribute in the portfolio websites show more accordion so screen readers will read that the accordion is expanded.