Updated accessibility and styles in the index.html, style.css and script.js files
Removed unnecessary comments in the index.html and style.css files.
Changed the semantic elements for the footer bottom copyright for better accessibility.
Added aria hidden to stop screen readers, reading out the icons from the footer bottom container in the index.html file.
Changed the style for the footer in the style.css file to make it look more professional and more accessible.
Added the dark mode functionality for the footer in the script.js file.