Updated accessibility, typography and styles in the index.html, style.css and script.js files
Added accessibility to any links that open in a new tab to be read out by screen readers to let users know what will happen when the link is clicked on in the index.html file.
Changed the section title's font family and capitalization, I also changed the position to make them stand out more in the index.html and style.css files.
Amended the footer to improve the design, readability and accessibility in the index.html and style.css files.
Added a new font family and removed the font family tomorrow light by putting it into a comment in case it is needed in the style.css file.
Added an ID selector to remove the contrast issue for accessibility in the style.css file.
Removed any unnecessary code in the script.js file.