Updated accessibility and styles in the index.html, style.css and script.js files
Removed any unnecessary code.
Updated accessibility for the links that open in a new tab, to ensure it was read in the correct order for screen readers in the index.html file.
Added text shadow, changed a margin unit for the portfolio websites title container class and added a font-size property for the section introduction container paragraphs in the style.css file.
Updated accessibility for the dark mode and hCaptcha modals in the script.js file.