Updated accessibility and styles in the index.html, style.css and script..js files
Removed any unnecessary code and comments.
Added and commented out, the portfolio website container for the future websites that will be added and tested the accessibility and ensured it  was working for the added container in the index.html file.
Changed the dark mode and hCaptcha modals, title, font family and overlay background to stop the accessibility contrast issues in the style.css file.
Removed any single quotes and added double quotes for readability in the script.js file.