Updated accessibility and styles in the index.html, style.css and script.js files
Changed the aria-label value to help with the redundant announcing bug in the index.html file.
Changed the view source code buttons href value, on the portfolio websites container to change it to the redirected link in the future in the index.html file.
Added a transition to the show more accordion container in the style.css and script.js files.
Made the commented out code more readable in the index.html, style.css and script.js files.
Fixed the hamburger menu dynamicAnnouncer accessibility contrast issue in DevTools.