Updated accessibility in the index.html and script.js files
Fixed aria-live, repeating text in the screen reader when the dark mode modal is closed in the script.js file.
Removed the testing message for the aria-label in the index.html file.
Changed the aria-live and aria-label values to make it clearer for the user in the script.js file.