Updated accessibility in the index.html and script.js files
Added a span element and an aria-hidden attribute to the submit buttons content in the index.html file.
Changed the function name for the hCaptcha modal when it closes in the script.js file.
Changed the script of the close hCaptcha modal, this was to ensure that the screen reader was not announcing any other element except for the submit button, I had to then ensure that it did not repeat the contact form submit buttons, .focus() property and then the aria-label straight after. To avoid redundant screen reader announcements, I had to  remove the aria label and add the aria-label with a blur event listener in the script.js file.
This was the only way that I could have the submit button focused when the modal closed and allow as little redundant screen reader announcements as possible.