Updated accessibility and styles in the index.html, style.css and script.js files
Added a comment to show where the skip to main content link will go for accessibility in the index.html file.
Added the focus pseudo classes to the links and buttons that have a hover affect, this improves accessibility and consistency in the style.css file.
Added the padding to the landing page title to make it more readable to users in the style.css file.
Removed any redundant code in the style.css file.
Removed any redundant code in the script.js file.
Added some commented out code and made a note as to what they do in the closeHcaptchaModal function, this is so that I can try to fix the issue in the future.