Updated accessibility in the index.html and script.js files
Added accessibility fixes to the skip to main content link, when the user clicks on the skip to main content link, the focus is set to the portfolio section immediately after.
Put some of the code into a comment, this was to fix redundant screen reader announcements. I need to make sure it does not conflict with anything else and make sure it is definitely not needed before removing the code in the index.html file.
Added the .scrollTo() JavaScript property to ensure that when users clicked on the skip to main content link, the page scrolled to the top of the section, rather than the first focusable element in the script.js file.